:recycle: [BE]Refactor: GET, DELETE, PATCH기능 구현수정 #42 #46 #47

+++ Server/be-springboot-starter-main/src/main/java/com/example/codestates/comment/controller/CommentController.java
@@ -1,10 +1,12 @@
 package com.example.codestates.comment.controller;
 
 import com.example.codestates.comment.dto.CommentDto;
+import com.example.codestates.comment.entity.Band;
 import com.example.codestates.comment.entity.Comment;
 import com.example.codestates.comment.mapper.CommentMapper;
-import com.example.codestates.comment.repository.CommentRepository;
+import com.example.codestates.comment.service.BandService;
 import com.example.codestates.comment.service.CommentService;
+import com.example.codestates.utils.UriCreator;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.http.ResponseEntity;
@@ -18,13 +20,14 @@ import java.util.List;
 @RequestMapping("/v1/comments")
 @Slf4j
 public class CommentController {
-    private final static String COMMENT_DEFAULT_URL ="/v1/comments";
+    private final static String COMMENT_DEFAULT_URL = "/v1/comments";
+    private final BandService bandService;
     private final CommentService commentService;
     private final CommentMapper mapper;
 
 
-
-    public CommentController(CommentService commentService, CommentMapper mapper) {
+    public CommentController(BandService bandService, CommentService commentService, CommentMapper mapper) {
+        this.bandService = bandService;
         this.commentService = commentService;
         this.mapper = mapper;
 
@@ -32,22 +35,40 @@ public class CommentController {
 
 
     @PostMapping(params = "band_id")
-    public ResponseEntity postComment(@RequestParam("band_id") long bandId,@RequestBody CommentDto.Post requestBody){
+    public ResponseEntity<Comment> postComment(@RequestParam("band_id") long bandId, @RequestBody CommentDto.Post requestBody) {
+        Band band = bandService.findById(bandId).orElse(null);
+        if(band == null){
+            return ResponseEntity.badRequest().build();
+        }
+
         Comment comment = mapper.commentPostDtoTocomment(requestBody);
-        Comment createdComment = commentService.createComment(comment);
-        URI location = UriCreator.createUri(COMMENT_DEFAULT_URL, createdComment.getCommentId());
-        return ResponseEntity.created(location).build();
+        comment.setBand(band);
+            Comment createdComment = commentService.createComment(bandId,comment);
+            URI location = UriCreator.createUri(COMMENT_DEFAULT_URL, createdComment.getCommentId());
+
+            return ResponseEntity.created(location).build();
+
     }
+
     @GetMapping(params = "band_id")
     public ResponseEntity getComment(@RequestParam("band_id") long bandId,
                                      @Positive @RequestParam int page,
                                      @Positive @RequestParam int size){
+        Band band = bandService.findById(bandId).orElse(null);
+        if(band == null){
+            return ResponseEntity.badRequest().build();
+        }
+
         Page<Comment> foundComment = commentService.findAll(page,size);
         List<Comment> comments = pageComment.getContent();
-        return ResponseEntity.ok(new MultiResponseDto(mapper.commentsToCommentResponseDtos(comments),pageComment))
+        return ResponseEntity.ok(new MultiResponseDto(mapper.commentsToCommentResponseDtos(comments),pageComment));
     }
     @DeleteMapping(params = {"band_id"},value = "/{comment_id}")
     public ResponseEntity deleteComment(@RequestParam("band_id") long bandId, @PathVariable("comment_id") long commentId){
+        Band band = bandService.findById(bandId).orElse(null);
+        if(band == null){
+            return ResponseEntity.badRequest().build();
+        }
         commentService.deleteComment(commentId);
         return ResponseEntity.noContent().build();
     }
@@ -55,6 +76,11 @@ public class CommentController {
     public ResponseEntity patchComment(@RequestParam("band_id") long bandId,
                                        @RequestParam("comment_id") long commentId,
                                       @RequestBody CommentDto.Patch requestBody ){
+        Band band = bandService.findById(bandId).orElse(null);
+        if(band == null){
+            return ResponseEntity.badRequest().build();
+        }
+
         requestBody.addCommentId(commentId);
         Comment updateComment = commentService.updateComment(mapper.commentPatchDtoTocomment(requestBody));
